requirements.txt
Flask==0.10.1
Jinja2==2.7.3
gunicorn==19.1.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
riotwatcher==3.1.1
urllib3==1.26.4
python-dotenv
